requirements.txt
# Core dependencies
PySide6>=6.5.0
cryptography>=38.0.0
requests>=2.28.0

# API dependencies
openai>=1.1.0
anthropic>=0.7.0
google-generativeai>=0.3.0
cohere>=4.30
mistralai>=0.0.7
ai21>=0.2.0
